experience_counter_djnago/exp_counter/requirements.txt
djangorestframework==3.11.2
Django==3.1.8
pytz==2021.3

